fixed the OR merge returning the wrong dtype, which broke route B in half precision
flex_attention returns lse in float32 no matter what dtype the inputs are, so
multiplying the half precision attention output by the softmax of the lse promoted
the merged result to float32. out_proj still has half weights, so the layer died
with "mat1 and mat2 must have the same dtype". Latent since step 2 and invisible
to the tests because they all run in float32, but training and benchmarking run in
half, so route B would have crashed on the first real forward pass.

The sum stays in float32, which is the better place to accumulate, and only the
result is cast back.

+++ src/hepattn/models/attention.py
@@ -110,7 +110,10 @@ def or_merge_lse(out: Tensor, lse: Tensor, n_hashes: int) -> Tensor:
     h = out.shape[1] // n_hashes
     out = out.unflatten(1, (h, n_hashes))  # (B, H, C, N, Dh)
     weights = torch.softmax(lse.unflatten(1, (h, n_hashes)), dim=2)  # (B, H, C, N) over C
-    return (out * weights.unsqueeze(-1)).sum(dim=2)
+    # flex returns lse in float32 whatever the input dtype, so the product promotes. Accumulate in
+    # that wider dtype (it is the numerically better place to sum) but hand back the dtype the rest
+    # of the layer uses, or out_proj sees float32 activations against half weights.
+    return (out * weights.unsqueeze(-1)).sum(dim=2).to(out.dtype)
 
 
 def unpad_for_flash_varlen(x: Tensor, kv_mask: Tensor) -> tuple[Tensor, Tensor, dict]:
